fix-use-server-declarations - rename _action to _actions
